requirements.txt
openai-whisper
torch
torchvision
torchaudio